+++ sklearn_genetic/callbacks/validations.py
@@ -1,6 +1,5 @@
-from typing import Callable
-
 from ..parameters import Metrics
+from .base import BaseCallback
 
 
 def check_stats(metric):
@@ -15,17 +14,18 @@ def check_callback(callback):
     Check if callback is a callable or a list of callables.
     """
     if callback is not None:
-        if isinstance(callback, Callable):
+        if isinstance(callback, BaseCallback):
             return [callback]
 
         elif isinstance(callback, list) and all(
-            [isinstance(c, Callable) for c in callback]
+            [isinstance(c, BaseCallback) for c in callback]
         ):
             return callback
 
         else:
             raise ValueError(
-                "callback should be either a callable or a list of callables."
+                "callback should be either a class or a list of classes with inheritance from "
+                "callbacks.base.BaseCallback"
             )
     else:
         return []